+++ jdplus-main-desktop/jdplus-sa-desktop-plugin/src/main/java/jdplus/sa/desktop/plugin/output/impl/CsvArraysOutputBuddy.java
@@ -34,6 +34,7 @@ import jdplus.toolkit.desktop.plugin.beans.BeanEditor;
 import jdplus.toolkit.desktop.plugin.beans.BeanHandler;
 import jdplus.toolkit.desktop.plugin.properties.NodePropertySetBuilder;
 import jdplus.toolkit.desktop.plugin.properties.PropertySheetDialogBuilder;
+import lombok.NonNull;
 import nbbrd.io.text.BooleanProperty;
 import nbbrd.io.text.Formatter;
 import nbbrd.io.text.Parser;
@@ -77,17 +78,17 @@ public final class CsvArraysOutputBuddy implements OutputFactoryBuddy, Configura
     }
 
     @Override
-    public Config getConfig() {
+    public @NonNull Config getConfig() {
         return configurator.getConfig(this);
     }
 
     @Override
-    public void setConfig(Config config) throws IllegalArgumentException {
+    public void setConfig(@NonNull Config config) throws IllegalArgumentException {
         configurator.setConfig(this, config);
     }
 
     @Override
-    public Config editConfig(Config config) throws IllegalArgumentException {
+    public @NonNull Config editConfig(@NonNull Config config) throws IllegalArgumentException {
         return configurator.editConfig(config);
     }